agument coarseningFactor is now a double and is used in the Multiple Pairwise Aggregation
Former-commit-id: 2800cb768acd643947f2233e1fefd832c1e02cf5

+++ src/Mesh/Mesh.h
@@ -303,7 +303,7 @@ public:
 		Utils::Warning("Impossible to export the solution to GMSH because this mesh does not come from GMSH.");
 	}
 
-	virtual void CoarsenMesh(CoarseningStrategy elemCoarseningStgy, FaceCoarseningStrategy faceCoarseningStgy, int coarseningFactor)
+	virtual void CoarsenMesh(CoarseningStrategy elemCoarseningStgy, FaceCoarseningStrategy faceCoarseningStgy, double coarseningFactor)
 	{
 		if (Utils::IsRefinementStrategy(elemCoarseningStgy))
 			return;
